+++ lib/src/metadata/mp3_metadata.dart
@@ -1,6 +1,5 @@
 // https://teslabs.com/openplayer/docs/docs/specs/id3v2.3.0%20-%20ID3.org.pdf
 
-import 'dart:collection';
 import 'dart:typed_data';
 
 import '../parsers/tag_parser.dart';