+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - FIX : read correcly a FLAC track. A mistake in the mask was missing the metadata
 - FIX : compute correctly the duration of a MP3 track with Variable Bit Rate (VBR)
+- add support for OPUS and OGG
 
 ## 0.0.1